+++ __tests__/fixtures.ts
@@ -1,8 +1,9 @@
 import fs from 'fs';
 import path from 'path';
 
-import { Account, ProviderInterface, RpcProvider, SequencerProvider, ec, json } from '../src';
-import { CompiledContract } from '../src/types';
+import { Account, ProviderInterface, RpcProvider, SequencerProvider, json } from '../src';
+import { CompiledContract, waitForTransactionOptions } from '../src/types';
+import { toHex } from '../src/utils/number';
 
 const readContract = (name: string): CompiledContract =>
   json.parse(
@@ -51,8 +52,11 @@ export const getTestProvider = (): ProviderInterface => {
   if (IS_LOCALHOST_DEVNET) {
     // accelerate the tests when running locally
     const originalWaitForTransaction = provider.waitForTransaction.bind(provider);
-    provider.waitForTransaction = (txHash: string, retryInterval?: number) => {
-      return originalWaitForTransaction(txHash, retryInterval || 1000);
+    provider.waitForTransaction = (
+      txHash: string,
+      { retryInterval }: waitForTransactionOptions = {}
+    ) => {
+      return originalWaitForTransaction(txHash, { retryInterval: retryInterval || 1000 });
     };
   }
 
@@ -76,7 +80,7 @@ export const getTestAccount = (provider: ProviderInterface) => {
     testAccountPrivateKey = DEFAULT_TEST_ACCOUNT_PRIVATE_KEY;
   }
 
-  return new Account(provider, testAccountAddress, ec.getKeyPair(testAccountPrivateKey));
+  return new Account(provider, toHex(testAccountAddress), testAccountPrivateKey);
 };
 
 const describeIf = (condition: boolean) => (condition ? describe : describe.skip);