+++ src/signer/default.ts
@@ -1,36 +1,31 @@
-import {
-  Abi,
-  Call,
-  DeclareSignerDetails,
-  DeployAccountSignerDetails,
-  InvocationsSignerDetails,
-  KeyPair,
-  Signature,
-} from '../types';
-import { genKeyPair, getStarkKey, sign } from '../utils/ellipticCurve';
+import { Abi, Call, DeclareSignerDetails, InvocationsSignerDetails, Signature } from '../types';
+import { DeployAccountSignerDetails } from '../types/signer';
+import { starkCurve } from '../utils/ec';
+import { buf2hex } from '../utils/encode';
 import {
   calculateDeclareTransactionHash,
   calculateDeployAccountTransactionHash,
   calculateTransactionHash,
 } from '../utils/hash';
+import { toHex } from '../utils/number';
 import { fromCallsToExecuteCalldata } from '../utils/transaction';
 import { TypedData, getMessageHash } from '../utils/typedData';
 import { SignerInterface } from './interface';
 
 export class Signer implements SignerInterface {
-  protected keyPair: KeyPair;
+  protected pk: Uint8Array | string;
 
-  constructor(keyPair: KeyPair = genKeyPair()) {
-    this.keyPair = keyPair;
+  constructor(pk: Uint8Array | string = starkCurve.utils.randomPrivateKey()) {
+    this.pk = pk instanceof Uint8Array ? buf2hex(pk) : toHex(pk);
   }
 
   public async getPubKey(): Promise<string> {
-    return getStarkKey(this.keyPair);
+    return starkCurve.getStarkKey(this.pk);
   }
 
   public async signMessage(typedData: TypedData, accountAddress: string): Promise<Signature> {
     const msgHash = getMessageHash(typedData, accountAddress);
-    return sign(this.keyPair, msgHash);
+    return starkCurve.sign(msgHash, this.pk);
   }
 
   public async signTransaction(
@@ -54,7 +49,7 @@ export class Signer implements SignerInterface {
       transactionsDetail.nonce
     );
 
-    return sign(this.keyPair, msgHash);
+    return starkCurve.sign(msgHash, this.pk);
   }
 
   public async signDeployAccountTransaction({
@@ -78,7 +73,7 @@ export class Signer implements SignerInterface {
       nonce
     );
 
-    return sign(this.keyPair, msgHash);
+    return starkCurve.sign(msgHash, this.pk);
   }
 
   public async signDeclareTransaction(
@@ -94,6 +89,6 @@ export class Signer implements SignerInterface {
       nonce
     );
 
-    return sign(this.keyPair, msgHash);
+    return starkCurve.sign(msgHash, this.pk);
   }
 }